Move legacy code blocks into seperate functions
