+++ crypto/evp/evp_enc.c
@@ -1,5 +1,5 @@
 /*
- * Copyright 1995-2019 The OpenSSL Project Authors. All Rights Reserved.
+ * Copyright 1995-2020 The OpenSSL Project Authors. All Rights Reserved.
  *
  * Licensed under the Apache License 2.0 (the "License").  You may not use
  * this file except in compliance with the License.  You can obtain a copy
@@ -7,27 +7,29 @@
  * https://www.openssl.org/source/license.html
  */
 
-#include <stdio.h>
-#include <assert.h>
-#include "internal/cryptlib.h"
+#include <string.h>
 #include <openssl/evp.h>
 #include <openssl/err.h>
 #include <openssl/rand.h>
-#include <openssl/rand_drbg.h>
-#include <openssl/engine.h>
-#include <openssl/params.h>
 #include <openssl/core_names.h>
 #include "crypto/evp.h"
 #include "internal/provider.h"
 #include "evp_local.h"
+#include "legacy_meth.h"
 
 int EVP_CIPHER_CTX_reset(EVP_CIPHER_CTX *ctx)
 {
     if (ctx == NULL)
         return 1;
 
-    if (ctx->cipher == NULL || ctx->cipher->prov == NULL)
-        goto legacy;
+#if !defined(FIPS_MODE)
+    {
+        int ret = 0;
+
+        if (legacy_evp_cipher_ctx_reset(ctx, &ret))
+            return ret;
+    }
+#endif
 
     if (ctx->provctx != NULL) {
         if (ctx->cipher->freectx != NULL)
@@ -38,23 +40,6 @@ int EVP_CIPHER_CTX_reset(EVP_CIPHER_CTX *ctx)
         EVP_CIPHER_free(ctx->fetched_cipher);
     memset(ctx, 0, sizeof(*ctx));
 
-    return 1;
-
-    /* TODO(3.0): Remove legacy code below */
- legacy:
-
-    if (ctx->cipher != NULL) {
-        if (ctx->cipher->cleanup && !ctx->cipher->cleanup(ctx))
-            return 0;
-        /* Cleanse cipher context data */
-        if (ctx->cipher_data && ctx->cipher->ctx_size)
-            OPENSSL_cleanse(ctx->cipher_data, ctx->cipher->ctx_size);
-    }
-    OPENSSL_free(ctx->cipher_data);
-#if !defined(OPENSSL_NO_ENGINE) && !defined(FIPS_MODE)
-    ENGINE_finish(ctx->engine);
-#endif
-    memset(ctx, 0, sizeof(*ctx));
     return 1;
 }
 
@@ -81,11 +66,6 @@ int EVP_CipherInit_ex(EVP_CIPHER_CTX *ctx, const EVP_CIPHER *cipher,
                       ENGINE *impl, const unsigned char *key,
                       const unsigned char *iv, int enc)
 {
-#if !defined(OPENSSL_NO_ENGINE) && !defined(FIPS_MODE)
-    ENGINE *tmpimpl = NULL;
-#endif
-    const EVP_CIPHER *tmpcipher;
-
     /*
      * enc == 1 means we are encrypting.
      * enc == 0 means we are decrypting.
@@ -104,208 +84,14 @@ int EVP_CipherInit_ex(EVP_CIPHER_CTX *ctx, const EVP_CIPHER *cipher,
         return 0;
     }
 
-    /* TODO(3.0): Legacy work around code below. Remove this */
+#if !defined(FIPS_MODE)
+    {
+        int ret = 0;
 
-#if !defined(OPENSSL_NO_ENGINE) && !defined(FIPS_MODE)
-    /*
-     * Whether it's nice or not, "Inits" can be used on "Final"'d contexts so
-     * this context may already have an ENGINE! Try to avoid releasing the
-     * previous handle, re-querying for an ENGINE, and having a
-     * reinitialisation, when it may all be unnecessary.
-     */
-    if (ctx->engine && ctx->cipher
-        && (cipher == NULL || cipher->nid == ctx->cipher->nid))
-        goto skip_to_init;
-
-    if (cipher != NULL && impl == NULL) {
-         /* Ask if an ENGINE is reserved for this job */
-        tmpimpl = ENGINE_get_cipher_engine(cipher->nid);
+        if (legacy_evp_cipher_init_ex(ctx, cipher, impl, key, iv, enc, &ret))
+            return ret;
     }
 #endif
-
-    /*
-     * If there are engines involved then we should use legacy handling for now.
-     */
-    if (ctx->engine != NULL
-#if !defined(OPENSSL_NO_ENGINE) && !defined(FIPS_MODE)
-            || tmpimpl != NULL
-#endif
-            || impl != NULL) {
-        if (ctx->cipher == ctx->fetched_cipher)
-            ctx->cipher = NULL;
-        EVP_CIPHER_free(ctx->fetched_cipher);
-        ctx->fetched_cipher = NULL;
-        goto legacy;
-    }
-
-    tmpcipher = (cipher == NULL) ? ctx->cipher : cipher;
-
-    if (tmpcipher->prov == NULL) {
-        switch(tmpcipher->nid) {
-        case NID_undef:
-        case NID_aes_256_ecb:
-        case NID_aes_192_ecb:
-        case NID_aes_128_ecb:
-        case NID_aes_256_cbc:
-        case NID_aes_192_cbc:
-        case NID_aes_128_cbc:
-        case NID_aes_256_ofb128:
-        case NID_aes_192_ofb128:
-        case NID_aes_128_ofb128:
-        case NID_aes_256_cfb128:
-        case NID_aes_192_cfb128:
-        case NID_aes_128_cfb128:
-        case NID_aes_256_cfb1:
-        case NID_aes_192_cfb1:
-        case NID_aes_128_cfb1:
-        case NID_aes_256_cfb8:
-        case NID_aes_192_cfb8:
-        case NID_aes_128_cfb8:
-        case NID_aes_256_ctr:
-        case NID_aes_192_ctr:
-        case NID_aes_128_ctr:
-        case NID_aes_128_xts:
-        case NID_aes_256_xts:
-        case NID_aes_256_ocb:
-        case NID_aes_192_ocb:
-        case NID_aes_128_ocb:
-        case NID_aes_256_gcm:
-        case NID_aes_192_gcm:
-        case NID_aes_128_gcm:
-        case NID_aes_256_siv:
-        case NID_aes_192_siv:
-        case NID_aes_128_siv:
-        case NID_aes_256_cbc_hmac_sha256:
-        case NID_aes_128_cbc_hmac_sha256:
-        case NID_aes_256_cbc_hmac_sha1:
-        case NID_aes_128_cbc_hmac_sha1:
-        case NID_id_aes256_wrap:
-        case NID_id_aes256_wrap_pad:
-        case NID_id_aes192_wrap:
-        case NID_id_aes192_wrap_pad:
-        case NID_id_aes128_wrap:
-        case NID_id_aes128_wrap_pad:
-        case NID_aria_256_gcm:
-        case NID_aria_192_gcm:
-        case NID_aria_128_gcm:
-        case NID_aes_256_ccm:
-        case NID_aes_192_ccm:
-        case NID_aes_128_ccm:
-        case NID_aria_256_ccm:
-        case NID_aria_192_ccm:
-        case NID_aria_128_ccm:
-        case NID_aria_256_ecb:
-        case NID_aria_192_ecb:
-        case NID_aria_128_ecb:
-        case NID_aria_256_cbc:
-        case NID_aria_192_cbc:
-        case NID_aria_128_cbc:
-        case NID_aria_256_ofb128:
-        case NID_aria_192_ofb128:
-        case NID_aria_128_ofb128:
-        case NID_aria_256_cfb128:
-        case NID_aria_192_cfb128:
-        case NID_aria_128_cfb128:
-        case NID_aria_256_cfb1:
-        case NID_aria_192_cfb1:
-        case NID_aria_128_cfb1:
-        case NID_aria_256_cfb8:
-        case NID_aria_192_cfb8:
-        case NID_aria_128_cfb8:
-        case NID_aria_256_ctr:
-        case NID_aria_192_ctr:
-        case NID_aria_128_ctr:
-        case NID_camellia_256_ecb:
-        case NID_camellia_192_ecb:
-        case NID_camellia_128_ecb:
-        case NID_camellia_256_cbc:
-        case NID_camellia_192_cbc:
-        case NID_camellia_128_cbc:
-        case NID_camellia_256_ofb128:
-        case NID_camellia_192_ofb128:
-        case NID_camellia_128_ofb128:
-        case NID_camellia_256_cfb128:
-        case NID_camellia_192_cfb128:
-        case NID_camellia_128_cfb128:
-        case NID_camellia_256_cfb1:
-        case NID_camellia_192_cfb1:
-        case NID_camellia_128_cfb1:
-        case NID_camellia_256_cfb8:
-        case NID_camellia_192_cfb8:
-        case NID_camellia_128_cfb8:
-        case NID_camellia_256_ctr:
-        case NID_camellia_192_ctr:
-        case NID_camellia_128_ctr:
-        case NID_des_ede3_cbc:
-        case NID_des_ede3_ecb:
-        case NID_des_ede3_ofb64:
-        case NID_des_ede3_cfb64:
-        case NID_des_ede3_cfb8:
-        case NID_des_ede3_cfb1:
-        case NID_des_ede_cbc:
-        case NID_des_ede_ecb:
-        case NID_des_ede_ofb64:
-        case NID_des_ede_cfb64:
-        case NID_desx_cbc:
-        case NID_des_cbc:
-        case NID_des_ecb:
-        case NID_des_cfb1:
-        case NID_des_cfb8:
-        case NID_des_cfb64:
-        case NID_des_ofb64:
-        case NID_id_smime_alg_CMS3DESwrap:
-        case NID_bf_cbc:
-        case NID_bf_ecb:
-        case NID_bf_cfb64:
-        case NID_bf_ofb64:
-        case NID_idea_cbc:
-        case NID_idea_ecb:
-        case NID_idea_cfb64:
-        case NID_idea_ofb64:
-        case NID_cast5_cbc:
-        case NID_cast5_ecb:
-        case NID_cast5_cfb64:
-        case NID_cast5_ofb64:
-        case NID_seed_cbc:
-        case NID_seed_ecb:
-        case NID_seed_cfb128:
-        case NID_seed_ofb128:
-        case NID_sm4_cbc:
-        case NID_sm4_ecb:
-        case NID_sm4_ctr:
-        case NID_sm4_cfb128:
-        case NID_sm4_ofb128:
-        case NID_rc4:
-        case NID_rc4_40:
-        case NID_rc5_cbc:
-        case NID_rc5_ecb:
-        case NID_rc5_cfb64:
-        case NID_rc5_ofb64:
-        case NID_rc2_cbc:
-        case NID_rc2_40_cbc:
-        case NID_rc2_64_cbc:
-        case NID_rc2_cfb64:
-        case NID_rc2_ofb64:
-        case NID_chacha20:
-        case NID_chacha20_poly1305:
-        case NID_rc4_hmac_md5:
-            break;
-        default:
-            goto legacy;
-        }
-    }
-
-    /*
-     * Ensure a context left lying around from last time is cleared
-     * (legacy code)
-     */
-    if (cipher != NULL && ctx->cipher != NULL) {
-        OPENSSL_clear_free(ctx->cipher_data, ctx->cipher->ctx_size);
-        ctx->cipher_data = NULL;
-    }
-
-
-    /* TODO(3.0): Start of non-legacy code below */
 
     /* Ensure a context left lying around from last time is cleared */
     if (cipher != NULL && ctx->cipher != NULL) {
@@ -387,138 +173,6 @@ int EVP_CipherInit_ex(EVP_CIPHER_CTX *ctx, const EVP_CIPHER *cipher,
                               iv,
                               iv == NULL ? 0
                                          : EVP_CIPHER_CTX_iv_length(ctx));
-
-    /* TODO(3.0): Remove legacy code below */
- legacy:
-
-    if (cipher != NULL) {
-        /*
-         * Ensure a context left lying around from last time is cleared (we
-         * previously attempted to avoid this if the same ENGINE and
-         * EVP_CIPHER could be used).
-         */
-        if (ctx->cipher) {
-            unsigned long flags = ctx->flags;
-            EVP_CIPHER_CTX_reset(ctx);
-            /* Restore encrypt and flags */
-            ctx->encrypt = enc;
-            ctx->flags = flags;
-        }
-#if !defined(OPENSSL_NO_ENGINE) && !defined(FIPS_MODE)
-        if (impl != NULL) {
-            if (!ENGINE_init(impl)) {
-                EVPerr(EVP_F_EVP_CIPHERINIT_EX, EVP_R_INITIALIZATION_ERROR);
-                return 0;
-            }
-        } else {
-            impl = tmpimpl;
-        }
-        if (impl != NULL) {
-            /* There's an ENGINE for this job ... (apparently) */
-            const EVP_CIPHER *c = ENGINE_get_cipher(impl, cipher->nid);
-
-            if (c == NULL) {
-                /*
-                 * One positive side-effect of US's export control history,
-                 * is that we should at least be able to avoid using US
-                 * misspellings of "initialisation"?
-                 */
-                EVPerr(EVP_F_EVP_CIPHERINIT_EX, EVP_R_INITIALIZATION_ERROR);
-                return 0;
-            }
-            /* We'll use the ENGINE's private cipher definition */
-            cipher = c;
-            /*
-             * Store the ENGINE functional reference so we know 'cipher' came
-             * from an ENGINE and we need to release it when done.
-             */
-            ctx->engine = impl;
-        } else {
-            ctx->engine = NULL;
-        }
-#endif
-
-        ctx->cipher = cipher;
-        if (ctx->cipher->ctx_size) {
-            ctx->cipher_data = OPENSSL_zalloc(ctx->cipher->ctx_size);
-            if (ctx->cipher_data == NULL) {
-                ctx->cipher = NULL;
-                EVPerr(EVP_F_EVP_CIPHERINIT_EX, ERR_R_MALLOC_FAILURE);
-                return 0;
-            }
-        } else {
-            ctx->cipher_data = NULL;
-        }
-        ctx->key_len = cipher->key_len;
-        /* Preserve wrap enable flag, zero everything else */
-        ctx->flags &= EVP_CIPHER_CTX_FLAG_WRAP_ALLOW;
-        if (ctx->cipher->flags & EVP_CIPH_CTRL_INIT) {
-            if (!EVP_CIPHER_CTX_ctrl(ctx, EVP_CTRL_INIT, 0, NULL)) {
-                ctx->cipher = NULL;
-                EVPerr(EVP_F_EVP_CIPHERINIT_EX, EVP_R_INITIALIZATION_ERROR);
-                return 0;
-            }
-        }
-    }
-#if !defined(OPENSSL_NO_ENGINE) && !defined(FIPS_MODE)
- skip_to_init:
-#endif
-    if (ctx->cipher == NULL)
-        return 0;
-
-    /* we assume block size is a power of 2 in *cryptUpdate */
-    OPENSSL_assert(ctx->cipher->block_size == 1
-                   || ctx->cipher->block_size == 8
-                   || ctx->cipher->block_size == 16);
-
-    if (!(ctx->flags & EVP_CIPHER_CTX_FLAG_WRAP_ALLOW)
-        && EVP_CIPHER_CTX_mode(ctx) == EVP_CIPH_WRAP_MODE) {
-        EVPerr(EVP_F_EVP_CIPHERINIT_EX, EVP_R_WRAP_MODE_NOT_ALLOWED);
-        return 0;
-    }
-
-    if (!(EVP_CIPHER_flags(EVP_CIPHER_CTX_cipher(ctx)) & EVP_CIPH_CUSTOM_IV)) {
-        switch (EVP_CIPHER_CTX_mode(ctx)) {
-
-        case EVP_CIPH_STREAM_CIPHER:
-        case EVP_CIPH_ECB_MODE:
-            break;
-
-        case EVP_CIPH_CFB_MODE:
-        case EVP_CIPH_OFB_MODE:
-
-            ctx->num = 0;
-            /* fall-through */
-
-        case EVP_CIPH_CBC_MODE:
-
-            OPENSSL_assert(EVP_CIPHER_CTX_iv_length(ctx) <=
-                           (int)sizeof(ctx->iv));
-            if (iv)
-                memcpy(ctx->oiv, iv, EVP_CIPHER_CTX_iv_length(ctx));
-            memcpy(ctx->iv, ctx->oiv, EVP_CIPHER_CTX_iv_length(ctx));
-            break;
-
-        case EVP_CIPH_CTR_MODE:
-            ctx->num = 0;
-            /* Don't reuse IV for CTR mode */
-            if (iv)
-                memcpy(ctx->iv, iv, EVP_CIPHER_CTX_iv_length(ctx));
-            break;
-
-        default:
-            return 0;
-        }
-    }
-
-    if (key || (ctx->cipher->flags & EVP_CIPH_ALWAYS_CALL_INIT)) {
-        if (!ctx->cipher->init(ctx, key, iv, enc))
-            return 0;
-    }
-    ctx->buf_len = 0;
-    ctx->final_used = 0;
-    ctx->block_mask = ctx->cipher->block_size - 1;
-    return 1;
 }
 
 int EVP_CipherUpdate(EVP_CIPHER_CTX *ctx, unsigned char *out, int *outl,
@@ -572,126 +226,10 @@ int EVP_DecryptInit_ex(EVP_CIPHER_CTX *ctx, const EVP_CIPHER *cipher,
     return EVP_CipherInit_ex(ctx, cipher, impl, key, iv, 0);
 }
 
-/*
- * According to the letter of standard difference between pointers
- * is specified to be valid only within same object. This makes
- * it formally challenging to determine if input and output buffers
- * are not partially overlapping with standard pointer arithmetic.
- */
-#ifdef PTRDIFF_T
-# undef PTRDIFF_T
-#endif
-#if defined(OPENSSL_SYS_VMS) && __INITIAL_POINTER_SIZE==64
-/*
- * Then we have VMS that distinguishes itself by adhering to
- * sizeof(size_t)==4 even in 64-bit builds, which means that
- * difference between two pointers might be truncated to 32 bits.
- * In the context one can even wonder how comparison for
- * equality is implemented. To be on the safe side we adhere to
- * PTRDIFF_T even for comparison for equality.
- */
-# define PTRDIFF_T uint64_t
-#else
-# define PTRDIFF_T size_t
-#endif
-
-int is_partially_overlapping(const void *ptr1, const void *ptr2, int len)
-{
-    PTRDIFF_T diff = (PTRDIFF_T)ptr1-(PTRDIFF_T)ptr2;
-    /*
-     * Check for partially overlapping buffers. [Binary logical
-     * operations are used instead of boolean to minimize number
-     * of conditional branches.]
-     */
-    int overlapped = (len > 0) & (diff != 0) & ((diff < (PTRDIFF_T)len) |
-                                                (diff > (0 - (PTRDIFF_T)len)));
-
-    return overlapped;
-}
-
-static int evp_EncryptDecryptUpdate(EVP_CIPHER_CTX *ctx,
-                                    unsigned char *out, int *outl,
-                                    const unsigned char *in, int inl)
-{
-    int i, j, bl, cmpl = inl;
-
-    if (EVP_CIPHER_CTX_test_flags(ctx, EVP_CIPH_FLAG_LENGTH_BITS))
-        cmpl = (cmpl + 7) / 8;
-
-    bl = ctx->cipher->block_size;
-
-    if (ctx->cipher->flags & EVP_CIPH_FLAG_CUSTOM_CIPHER) {
-        /* If block size > 1 then the cipher will have to do this check */
-        if (bl == 1 && is_partially_overlapping(out, in, cmpl)) {
-            EVPerr(EVP_F_EVP_ENCRYPTDECRYPTUPDATE, EVP_R_PARTIALLY_OVERLAPPING);
-            return 0;
-        }
-
-        i = ctx->cipher->do_cipher(ctx, out, in, inl);
-        if (i < 0)
-            return 0;
-        else
-            *outl = i;
-        return 1;
-    }
-
-    if (inl <= 0) {
-        *outl = 0;
-        return inl == 0;
-    }
-    if (is_partially_overlapping(out + ctx->buf_len, in, cmpl)) {
-        EVPerr(EVP_F_EVP_ENCRYPTDECRYPTUPDATE, EVP_R_PARTIALLY_OVERLAPPING);
-        return 0;
-    }
-
-    if (ctx->buf_len == 0 && (inl & (ctx->block_mask)) == 0) {
-        if (ctx->cipher->do_cipher(ctx, out, in, inl)) {
-            *outl = inl;
-            return 1;
-        } else {
-            *outl = 0;
-            return 0;
-        }
-    }
-    i = ctx->buf_len;
-    OPENSSL_assert(bl <= (int)sizeof(ctx->buf));
-    if (i != 0) {
-        if (bl - i > inl) {
-            memcpy(&(ctx->buf[i]), in, inl);
-            ctx->buf_len += inl;
-            *outl = 0;
-            return 1;
-        } else {
-            j = bl - i;
-            memcpy(&(ctx->buf[i]), in, j);
-            inl -= j;
-            in += j;
-            if (!ctx->cipher->do_cipher(ctx, out, ctx->buf, bl))
-                return 0;
-            out += bl;
-            *outl = bl;
-        }
-    } else
-        *outl = 0;
-    i = inl & (bl - 1);
-    inl -= i;
-    if (inl > 0) {
-        if (!ctx->cipher->do_cipher(ctx, out, in, inl))
-            return 0;
-        *outl += inl;
-    }
-
-    if (i != 0)
-        memcpy(ctx->buf, &(in[inl]), i);
-    ctx->buf_len = i;
-    return 1;
-}
-
-
 int EVP_EncryptUpdate(EVP_CIPHER_CTX *ctx, unsigned char *out, int *outl,
                       const unsigned char *in, int inl)
 {
-    int ret;
+    int ret = 0;
     size_t soutl;
     int blocksize;
 
@@ -706,8 +244,10 @@ int EVP_EncryptUpdate(EVP_CIPHER_CTX *ctx, unsigned char *out, int *outl,
         return 0;
     }
 
-    if (ctx->cipher->prov == NULL)
-        goto legacy;
+#if !defined(FIPS_MODE)
+    if (legacy_evp_encrypt_update(ctx, out, outl, in, inl, &ret))
+        return ret;
+#endif
 
     blocksize = EVP_CIPHER_CTX_block_size(ctx);
 
@@ -726,26 +266,17 @@ int EVP_EncryptUpdate(EVP_CIPHER_CTX *ctx, unsigned char *out, int *outl,
         }
         *outl = soutl;
     }
-
     return ret;
-
-    /* TODO(3.0): Remove legacy code below */
- legacy:
-
-    return evp_EncryptDecryptUpdate(ctx, out, outl, in, inl);
 }
 
 int EVP_EncryptFinal(EVP_CIPHER_CTX *ctx, unsigned char *out, int *outl)
 {
-    int ret;
-    ret = EVP_EncryptFinal_ex(ctx, out, outl);
-    return ret;
+    return EVP_EncryptFinal_ex(ctx, out, outl);
 }
 
 int EVP_EncryptFinal_ex(EVP_CIPHER_CTX *ctx, unsigned char *out, int *outl)
 {
-    int n, ret;
-    unsigned int i, b, bl;
+    int ret = 0;
     size_t soutl;
     int blocksize;
 
@@ -759,8 +290,11 @@ int EVP_EncryptFinal_ex(EVP_CIPHER_CTX *ctx, unsigned char *out, int *outl)
         EVPerr(EVP_F_EVP_ENCRYPTFINAL_EX, EVP_R_NO_CIPHER_SET);
         return 0;
     }
-    if (ctx->cipher->prov == NULL)
-        goto legacy;
+
+#if !defined(FIPS_MODE)
+    if (legacy_evp_encrypt_final_ex(ctx, out, outl, &ret))
+        return ret;
+#endif
 
     blocksize = EVP_CIPHER_CTX_block_size(ctx);
 
@@ -779,54 +313,13 @@ int EVP_EncryptFinal_ex(EVP_CIPHER_CTX *ctx, unsigned char *out, int *outl)
         }
         *outl = soutl;
     }
-
-    return ret;
-
-    /* TODO(3.0): Remove legacy code below */
- legacy:
-
-    if (ctx->cipher->flags & EVP_CIPH_FLAG_CUSTOM_CIPHER) {
-        ret = ctx->cipher->do_cipher(ctx, out, NULL, 0);
-        if (ret < 0)
-            return 0;
-        else
-            *outl = ret;
-        return 1;
-    }
-
-    b = ctx->cipher->block_size;
-    OPENSSL_assert(b <= sizeof(ctx->buf));
-    if (b == 1) {
-        *outl = 0;
-        return 1;
-    }
-    bl = ctx->buf_len;
-    if (ctx->flags & EVP_CIPH_NO_PADDING) {
-        if (bl) {
-            EVPerr(EVP_F_EVP_ENCRYPTFINAL_EX,
-                   EVP_R_DATA_NOT_MULTIPLE_OF_BLOCK_LENGTH);
-            return 0;
-        }
-        *outl = 0;
-        return 1;
-    }
-
-    n = b - bl;
-    for (i = bl; i < b; i++)
-        ctx->buf[i] = n;
-    ret = ctx->cipher->do_cipher(ctx, out, ctx->buf, b);
-
-    if (ret)
-        *outl = b;
-
     return ret;
 }
 
 int EVP_DecryptUpdate(EVP_CIPHER_CTX *ctx, unsigned char *out, int *outl,
                       const unsigned char *in, int inl)
 {
-    int fix_len, cmpl = inl, ret;
-    unsigned int b;
+    int ret = 0;
     size_t soutl;
     int blocksize;
 
@@ -840,8 +333,11 @@ int EVP_DecryptUpdate(EVP_CIPHER_CTX *ctx, unsigned char *out, int *outl,
         EVPerr(EVP_F_EVP_DECRYPTUPDATE, EVP_R_NO_CIPHER_SET);
         return 0;
     }
-    if (ctx->cipher->prov == NULL)
-        goto legacy;
+
+#if !defined(FIPS_MODE)
+    if (legacy_evp_decrypt_update(ctx, out, outl, in, inl, &ret))
+        return ret;
+#endif
 
     blocksize = EVP_CIPHER_CTX_block_size(ctx);
 
@@ -862,86 +358,17 @@ int EVP_DecryptUpdate(EVP_CIPHER_CTX *ctx, unsigned char *out, int *outl,
     }
 
     return ret;
-
-    /* TODO(3.0): Remove legacy code below */
- legacy:
-
-    b = ctx->cipher->block_size;
-
-    if (EVP_CIPHER_CTX_test_flags(ctx, EVP_CIPH_FLAG_LENGTH_BITS))
-        cmpl = (cmpl + 7) / 8;
-
-    if (ctx->cipher->flags & EVP_CIPH_FLAG_CUSTOM_CIPHER) {
-        if (b == 1 && is_partially_overlapping(out, in, cmpl)) {
-            EVPerr(EVP_F_EVP_DECRYPTUPDATE, EVP_R_PARTIALLY_OVERLAPPING);
-            return 0;
-        }
-
-        fix_len = ctx->cipher->do_cipher(ctx, out, in, inl);
-        if (fix_len < 0) {
-            *outl = 0;
-            return 0;
-        } else
-            *outl = fix_len;
-        return 1;
-    }
-
-    if (inl <= 0) {
-        *outl = 0;
-        return inl == 0;
-    }
-
-    if (ctx->flags & EVP_CIPH_NO_PADDING)
-        return evp_EncryptDecryptUpdate(ctx, out, outl, in, inl);
-
-    OPENSSL_assert(b <= sizeof(ctx->final));
-
-    if (ctx->final_used) {
-        /* see comment about PTRDIFF_T comparison above */
-        if (((PTRDIFF_T)out == (PTRDIFF_T)in)
-            || is_partially_overlapping(out, in, b)) {
-            EVPerr(EVP_F_EVP_DECRYPTUPDATE, EVP_R_PARTIALLY_OVERLAPPING);
-            return 0;
-        }
-        memcpy(out, ctx->final, b);
-        out += b;
-        fix_len = 1;
-    } else
-        fix_len = 0;
-
-    if (!evp_EncryptDecryptUpdate(ctx, out, outl, in, inl))
-        return 0;
-
-    /*
-     * if we have 'decrypted' a multiple of block size, make sure we have a
-     * copy of this last block
-     */
-    if (b > 1 && !ctx->buf_len) {
-        *outl -= b;
-        ctx->final_used = 1;
-        memcpy(ctx->final, &out[*outl], b);
-    } else
-        ctx->final_used = 0;
-
-    if (fix_len)
-        *outl += b;
-
-    return 1;
 }
 
 int EVP_DecryptFinal(EVP_CIPHER_CTX *ctx, unsigned char *out, int *outl)
 {
-    int ret;
-    ret = EVP_DecryptFinal_ex(ctx, out, outl);
-    return ret;
+    return EVP_DecryptFinal_ex(ctx, out, outl);
 }
 
 int EVP_DecryptFinal_ex(EVP_CIPHER_CTX *ctx, unsigned char *out, int *outl)
 {
-    int i, n;
-    unsigned int b;
     size_t soutl;
-    int ret;
+    int ret = 0;
     int blocksize;
 
     /* Prevent accidental use of encryption context when decrypting */
@@ -955,8 +382,10 @@ int EVP_DecryptFinal_ex(EVP_CIPHER_CTX *ctx, unsigned char *out, int *outl)
         return 0;
     }
 
-    if (ctx->cipher->prov == NULL)
-        goto legacy;
+#if !defined(FIPS_MODE)
+    if (legacy_evp_decrypt_final_ex(ctx, out, outl, &ret))
+        return ret;
+#endif
 
     blocksize = EVP_CIPHER_CTX_block_size(ctx);
 
@@ -975,67 +404,17 @@ int EVP_DecryptFinal_ex(EVP_CIPHER_CTX *ctx, unsigned char *out, int *outl)
         }
         *outl = soutl;
     }
-
     return ret;
-
-    /* TODO(3.0): Remove legacy code below */
- legacy:
-
-    *outl = 0;
-    if (ctx->cipher->flags & EVP_CIPH_FLAG_CUSTOM_CIPHER) {
-        i = ctx->cipher->do_cipher(ctx, out, NULL, 0);
-        if (i < 0)
-            return 0;
-        else
-            *outl = i;
-        return 1;
-    }
-
-    b = ctx->cipher->block_size;
-    if (ctx->flags & EVP_CIPH_NO_PADDING) {
-        if (ctx->buf_len) {
-            EVPerr(EVP_F_EVP_DECRYPTFINAL_EX,
-                   EVP_R_DATA_NOT_MULTIPLE_OF_BLOCK_LENGTH);
-            return 0;
-        }
-        *outl = 0;
-        return 1;
-    }
-    if (b > 1) {
-        if (ctx->buf_len || !ctx->final_used) {
-            EVPerr(EVP_F_EVP_DECRYPTFINAL_EX, EVP_R_WRONG_FINAL_BLOCK_LENGTH);
-            return 0;
-        }
-        OPENSSL_assert(b <= sizeof(ctx->final));
-
-        /*
-         * The following assumes that the ciphertext has been authenticated.
-         * Otherwise it provides a padding oracle.
-         */
-        n = ctx->final[b - 1];
-        if (n == 0 || n > (int)b) {
-            EVPerr(EVP_F_EVP_DECRYPTFINAL_EX, EVP_R_BAD_DECRYPT);
-            return 0;
-        }
-        for (i = 0; i < n; i++) {
-            if (ctx->final[--b] != n) {
-                EVPerr(EVP_F_EVP_DECRYPTFINAL_EX, EVP_R_BAD_DECRYPT);
-                return 0;
-            }
-        }
-        n = ctx->cipher->block_size - n;
-        for (i = 0; i < n; i++)
-            out[i] = ctx->final[i];
-        *outl = n;
-    } else
-        *outl = 0;
-    return 1;
 }
 
 int EVP_CIPHER_CTX_set_key_length(EVP_CIPHER_CTX *c, int keylen)
 {
-    if (c->cipher->prov != NULL) {
-        int ok;
+    int ret = 0;
+#if !defined(FIPS_MODE)
+    if (legacy_evp_cipher_ctx_set_key_length(c, keylen, &ret))
+        return ret;
+#endif
+    {
         OSSL_PARAM params[2] = { OSSL_PARAM_END, OSSL_PARAM_END };
         size_t len = keylen;
 
@@ -1048,27 +427,10 @@ int EVP_CIPHER_CTX_set_key_length(EVP_CIPHER_CTX *c, int keylen)
             return 0;
 
         params[0] = OSSL_PARAM_construct_size_t(OSSL_CIPHER_PARAM_KEYLEN, &len);
-        ok = evp_do_ciph_ctx_setparams(c->cipher, c->provctx, params);
+        ret = evp_do_ciph_ctx_setparams(c->cipher, c->provctx, params);
 
-        return ok > 0 ? 1 : 0;
+        return ret > 0 ? 1 : 0;
     }
-
-    /* TODO(3.0) legacy code follows */
-
-    /*
-     * Note there have never been any built-in ciphers that define this flag
-     * since it was first introduced.
-     */
-    if (c->cipher->flags & EVP_CIPH_CUSTOM_KEY_LENGTH)
-        return EVP_CIPHER_CTX_ctrl(c, EVP_CTRL_SET_KEY_LENGTH, keylen, NULL);
-    if (EVP_CIPHER_CTX_key_length(c) == keylen)
-        return 1;
-    if ((keylen > 0) && (c->cipher->flags & EVP_CIPH_VARIABLE_LENGTH)) {
-        c->key_len = keylen;
-        return 1;
-    }
-    EVPerr(EVP_F_EVP_CIPHER_CTX_SET_KEY_LENGTH, EVP_R_INVALID_KEY_LENGTH);
-    return 0;
 }
 
 int EVP_CIPHER_CTX_set_padding(EVP_CIPHER_CTX *ctx, int pad)
@@ -1102,10 +464,10 @@ int EVP_CIPHER_CTX_ctrl(EVP_CIPHER_CTX *ctx, int type, int arg, void *ptr)
         EVPerr(EVP_F_EVP_CIPHER_CTX_CTRL, EVP_R_NO_CIPHER_SET);
         return 0;
     }
-
-    if (ctx->cipher->prov == NULL)
-        goto legacy;
-
+#if !defined(FIPS_MODE)
+    if (legacy_evp_cipher_ctx_ctrl(ctx, type, arg, ptr, &ret))
+        return ret;
+#endif
     switch (type) {
     case EVP_CTRL_SET_KEY_LENGTH:
         params[0] = OSSL_PARAM_construct_size_t(OSSL_CIPHER_PARAM_KEYLEN, &sz);
@@ -1273,18 +635,7 @@ int EVP_CIPHER_CTX_ctrl(EVP_CIPHER_CTX *ctx, int type, int arg, void *ptr)
         ret = evp_do_ciph_ctx_setparams(ctx->cipher, ctx->provctx, params);
     else
         ret = evp_do_ciph_ctx_getparams(ctx->cipher, ctx->provctx, params);
-    goto end;
-
-/* TODO(3.0): Remove legacy code below */
-legacy:
-    if (ctx->cipher->ctrl == NULL) {
-        EVPerr(EVP_F_EVP_CIPHER_CTX_CTRL, EVP_R_CTRL_NOT_IMPLEMENTED);
-        return 0;
-    }
-
-    ret = ctx->cipher->ctrl(ctx, type, arg, ptr);
-
- end:
+end:
     if (ret == EVP_CTRL_RET_UNSUPPORTED) {
         EVPerr(EVP_F_EVP_CIPHER_CTX_CTRL,
                EVP_R_CTRL_OPERATION_NOT_IMPLEMENTED);
@@ -1361,8 +712,14 @@ int EVP_CIPHER_CTX_copy(EVP_CIPHER_CTX *out, const EVP_CIPHER_CTX *in)
         return 0;
     }
 
-    if (in->cipher->prov == NULL)
-        goto legacy;
+#if !defined(FIPS_MODE)
+    {
+        int ret = 0;
+
+        if (legacy_evp_cipher_ctx_copy(out, in, &ret))
+            return ret;
+    }
+#endif
 
     if (in->cipher->dupctx == NULL) {
         EVPerr(EVP_F_EVP_CIPHER_CTX_COPY, EVP_R_NOT_ABLE_TO_COPY_CTX);
@@ -1386,38 +743,6 @@ int EVP_CIPHER_CTX_copy(EVP_CIPHER_CTX *out, const EVP_CIPHER_CTX *in)
     }
 
     return 1;
-
-    /* TODO(3.0): Remove legacy code below */
- legacy:
-
-#if !defined(OPENSSL_NO_ENGINE) && !defined(FIPS_MODE)
-    /* Make sure it's safe to copy a cipher context using an ENGINE */
-    if (in->engine && !ENGINE_init(in->engine)) {
-        EVPerr(EVP_F_EVP_CIPHER_CTX_COPY, ERR_R_ENGINE_LIB);
-        return 0;
-    }
-#endif
-
-    EVP_CIPHER_CTX_reset(out);
-    memcpy(out, in, sizeof(*out));
-
-    if (in->cipher_data && in->cipher->ctx_size) {
-        out->cipher_data = OPENSSL_malloc(in->cipher->ctx_size);
-        if (out->cipher_data == NULL) {
-            out->cipher = NULL;
-            EVPerr(EVP_F_EVP_CIPHER_CTX_COPY, ERR_R_MALLOC_FAILURE);
-            return 0;
-        }
-        memcpy(out->cipher_data, in->cipher_data, in->cipher->ctx_size);
-    }
-
-    if (in->cipher->flags & EVP_CIPH_CUSTOM_COPY)
-        if (!in->cipher->ctrl((EVP_CIPHER_CTX *)in, EVP_CTRL_COPY, 0, out)) {
-            out->cipher = NULL;
-            EVPerr(EVP_F_EVP_CIPHER_CTX_COPY, EVP_R_INITIALIZATION_ERROR);
-            return 0;
-        }
-    return 1;
 }
 
 EVP_CIPHER *evp_cipher_new(void)
@@ -1435,37 +760,6 @@ EVP_CIPHER *evp_cipher_new(void)
     return cipher;
 }
 
-/*
- * FIPS module note: since internal fetches will be entirely
- * provider based, we know that none of its code depends on legacy
- * NIDs or any functionality that use them.
- */
-#ifndef FIPS_MODE
-/* TODO(3.x) get rid of the need for legacy NIDs */
-static void set_legacy_nid(const char *name, void *vlegacy_nid)
-{
-    int nid;
-    int *legacy_nid = vlegacy_nid;
-    /*
-     * We use lowest level function to get the associated method, because
-     * higher level functions such as EVP_get_cipherbyname() have changed
-     * to look at providers too.
-     */
-    const void *legacy_method = OBJ_NAME_get(name, OBJ_NAME_TYPE_CIPHER_METH);
-
-    if (*legacy_nid == -1)       /* We found a clash already */
-        return;
-    if (legacy_method == NULL)
-        return;
-    nid = EVP_CIPHER_nid(legacy_method);
-    if (*legacy_nid != NID_undef && *legacy_nid != nid) {
-        *legacy_nid = -1;
-        return;
-    }
-    *legacy_nid = nid;
-}
-#endif
-
 static void *evp_cipher_from_dispatch(const int name_id,
                                       const OSSL_DISPATCH *fns,
                                       OSSL_PROVIDER *prov)
@@ -1481,7 +775,7 @@ static void *evp_cipher_from_dispatch(const int name_id,
 #ifndef FIPS_MODE
     /* TODO(3.x) get rid of the need for legacy NIDs */
     cipher->nid = NID_undef;
-    evp_names_do_all(prov, name_id, set_legacy_nid, &cipher->nid);
+    evp_names_do_all(prov, name_id, legacy_evp_cipher_set_nid, &cipher->nid);
     if (cipher->nid == -1) {
         ERR_raise(ERR_LIB_EVP, ERR_R_INTERNAL_ERROR);
         EVP_CIPHER_free(cipher);
