+++ crypto/evp/legacy_meth.h
@@ -38,3 +38,39 @@ static int nm##_final(EVP_MD_CTX *ctx, unsigned char *md)                      \
 
 #define LEGACY_EVP_MD_METH_TABLE(init, update, final, ctrl, blksz)             \
     init, update, final, NULL, NULL, blksz, 0, ctrl
+
+void legacy_evp_cipher_set_nid(const char *name, void *vlegacy_nid);
+int legacy_evp_cipher_ctx_reset(EVP_CIPHER_CTX *ctx, int *ret);
+int legacy_evp_cipher_ctx_set_key_length(EVP_CIPHER_CTX *c, int keylen,
+                                         int *ret);
+int legacy_evp_cipher_ctx_ctrl(EVP_CIPHER_CTX *ctx, int type, int arg,
+                               void *ptr, int *ret);
+int legacy_evp_cipher_ctx_copy(EVP_CIPHER_CTX *out, const EVP_CIPHER_CTX *in,
+                               int *ret);
+
+int legacy_evp_cipher_init_ex(EVP_CIPHER_CTX *ctx, const EVP_CIPHER *cipher,
+                              ENGINE *impl, const unsigned char *key,
+                              const unsigned char *iv, int enc, int *ret);
+int legacy_evp_encrypt_update(EVP_CIPHER_CTX *ctx, unsigned char *out,
+                              int *outl, const unsigned char *in, int inl,
+                              int *ret);
+int legacy_evp_encrypt_final_ex(EVP_CIPHER_CTX *ctx, unsigned char *out,
+                                int *outl, int *res);
+int legacy_evp_decrypt_update(EVP_CIPHER_CTX *ctx, unsigned char *out, int *outl,
+                              const unsigned char *in, int inl, int *res);
+int legacy_evp_decrypt_final_ex(EVP_CIPHER_CTX *ctx, unsigned char *out,
+                                int *outl, int *ret);
+
+void legacy_evp_digest_set_nid(const char *name, void *vlegacy_nid);
+void legacy_evp_md_ctx_reset(EVP_MD_CTX *ctx);
+int legacy_evp_md_ctx_copy_ex(EVP_MD_CTX *out, const EVP_MD_CTX *in, int *ret);
+int legacy_evp_md_ctx_ctrl(EVP_MD_CTX *ctx, int cmd, int p1, void *p2, int *ret);
+
+int legacy_evp_digest_init_ex(EVP_MD_CTX *ctx, const EVP_MD *type, ENGINE *impl,
+                              int *ret);
+int legacy_evp_digest_update(EVP_MD_CTX *ctx, const void *data, size_t count,
+                             int *ret);
+int legacy_evp_digest_final_ex(EVP_MD_CTX *ctx, unsigned char *md,
+                               unsigned int *isize, int *ret);
+int legacy_evp_digest_final_xof(EVP_MD_CTX *ctx, unsigned char *md, size_t size,
+                                int *ret);
